Rename process source directories to match launch names.
src/proxy, src/reverse-proxy, and src/automation become qemu-server,
qemu-reverse-proxy, and automation-server. Tests, wrappers, identifiers,
and npm scripts follow the same paths.
