+++ package.json
@@ -11,9 +11,9 @@
     "client": "node --experimental-strip-types src/client/main.ts",
     "ctrl": "node --experimental-strip-types src/ctrl/main.ts",
     "session": "node --experimental-strip-types src/session/main.ts",
-    "qemu-server": "node --experimental-strip-types --import ./src/observability/instrument.ts src/proxy/main.ts",
-    "qemu-reverse-proxy": "node --experimental-strip-types --import ./src/observability/instrument.ts src/reverse-proxy/main.ts",
-    "automation-server": "node --experimental-strip-types --import ./src/observability/instrument.ts src/automation/main.ts",
+    "qemu-server": "node --experimental-strip-types --import ./src/observability/instrument.ts src/qemu-server/main.ts",
+    "qemu-reverse-proxy": "node --experimental-strip-types --import ./src/observability/instrument.ts src/qemu-reverse-proxy/main.ts",
+    "automation-server": "node --experimental-strip-types --import ./src/observability/instrument.ts src/automation-server/main.ts",
     "dev": "wrangler dev --remote",
     "check:lint": "oxlint --deny-warnings .",
     "check:format": "oxfmt --check .",
