+++ src/automation-server/command.ts
