+++ src/automation-server/signature.ts
