+++ src/qemu-reverse-proxy/handlers.ts
@@ -1,7 +1,7 @@
 import { Effect, Layer } from "effect";
 import { HttpApiBuilder } from "effect/unstable/httpapi";
-import * as ProxyHandlers from "../proxy/handlers.ts";
-import * as Middleware from "../proxy/middleware.ts";
+import * as ProxyHandlers from "../qemu-server/handlers.ts";
+import * as Middleware from "../qemu-server/middleware.ts";
 import * as Api from "../shared/api.ts";
 import * as Contract from "../shared/contract.ts";
 import * as Router from "./router.ts";
