+++ src/qemu-server/command.ts
