+++ src/qemu-server/middleware.ts
