+++ test/automation-server/command.unit.test.ts
@@ -16,7 +16,7 @@ import { TestConsole } from "effect/testing";
 import { Command } from "effect/unstable/cli";
 import { HttpServerError } from "effect/unstable/http";
 import { ChildProcessSpawner } from "effect/unstable/process";
-import * as AutomationCommand from "../../src/automation/command.ts";
+import * as AutomationCommand from "../../src/automation-server/command.ts";
 import * as Client from "../../src/db/client.ts";
 import * as Api from "../../src/shared/api.ts";
 import * as Errors from "../../src/shared/errors.ts";
